Move CameraApi to MethodCallHandlerImpl
